refactor(compile): Compile now sourced with `luafile`
Precompiled files are now sourced with the command `luafile`. This lets
the precompiled files live outside of the lua runtimepath. This means
that the default location of the precompiled files can be the system's
cache directory instead of neovim's config/lua folder.

Since the file is now executed by sourcing, there is no way to passing
which style to use. This meant that the precompiled file needed to be
broken up into one file for each style.